AllocationOpt

Documentation for AllocationOpt.

AllocationOpt.optimize_indexer

AllocationOpt.optimize_indexer — Method

optimizeindexer(;id::String, grtgas::Float64, alloclifetime::Int64, whitelist::Union{Nothing,Vector{String}}, blacklist::Union{Nothing,Vector{String}}, alloclifetimethreshold::Int64, preference_ratio::Float64)::DataFrame

Arguments

id::String: indexer address

grtgas::Float64: the price you would want pay for each allocation transaction (grtgas * 1 for open, grtgas * 1 for close, grtgas * 0.3 for claim). Higher setting will lead to smaller set of allocations.

alloc_lifetime::Int64: the frequency to renew a specific allocation. Smaller alloction lifetime allows less time to accumulate indexing rewards, thus lead to smaller set of allocations.

whitelist::Union{Nothing,Vector{String}}: when set, optimizer will only consider these subgraphs (Qm...), and you cannot set blacklist.

blacklist::Union{Nothing,Vector{String}}: when set, optimizer will not consider these subgraphs (Qm...), and you cannot set whitelist.

alloc_lifetime_threshold::Int64: determine which current allocations should be re-considered during optimization. With higher threshold, the optimizer blacklist more subgrpahs. If you set this to 0, all current allocations will be considered in optimization.

preference_ratio::Float64: the ratio between reallocating or keep open the current allocations. If you set to 1.0, optimizer simply takes the higher reward. If you set this to >1.0, you prefer to reallocate, if you set this to <1.0, you prefer to keep current allocation until it expires.