AllocationOpt

Documentation for AllocationOpt.

AllocationOpt.optimize_indexer

AllocationOpt.optimize_indexer — Method

function optimize_indexer(id, whitelist, blacklist, pinnedlist, frozenlist)

Arguments

id::AbstractString: The id of the indexer to optimise.

whitelist::Vector{AbstractString}: Subgraph deployment IPFS hashes included in this list will be considered for, but not guaranteed allocation.

blacklist::Vector{AbstractString}: Subgraph deployment IPFS hashes included in this list will not be considered, and will be suggested to close if there's an existing allocation.

pinnedlist::Vector{AbstractString}: Subgraph deployment IPFS hashes included in this list will be guaranteed allocation. Currently unsupported.

frozenlist::Vector{AbstractString}: Subgraph deployment IPFS hashes included in this list will not be considered during optimisation. Any allocations you have on these subgraphs deployments will remain.

Examples

julia> using AllocationOpt
julia> optimize_indexer("0x6ac85b9d834b51b14a7b0ed849bb5199e04c05c5", String["QmP4oSiQ7Wc4JTFk86m2JxGvR912NyBbxJnEdZawkYLTk4"], String[], String[], String[])
1-element Vector{Tuple{String, Float64}}:
 ("QmP4oSiQ7Wc4JTFk86m2JxGvR912NyBbxJnEdZawkYLTk4", 5.539482411224138e6)
julia> optimize_indexer("foo", String["QmP4oSiQ7Wc4JTFk86m2JxGvR912NyBbxJnEdZawkYLTk4"], String[], String[], String[])
ERROR: AllocationOpt.UnknownIndexerError()
julia> optimize_indexer("0x6ac85b9d834b51b14a7b0ed849bb5199e04c05c5", String["foo"], String[], String[], String[])
ERROR: AllocationOpt.BadSubgraphIpfsHashError()
Stacktrace:
 [1] optimize_indexer(id::String, whitelist::Vector{String}, blacklist::Vector{String}, pinnedlist::Vector{String}, frozenlist::Vector{String})