AllocationOpt

Documentation for AllocationOpt.

AllocationOpt is a library for optimising how an indexer should allocate its stake in The Graph Protocol.

Important: You must run this on a computer that supports 64-bit operations. 32-bit doesn't have enough precision to solve the optimisation problem.

Installation

Enter the julia repl. On linux machines, this is as simple as running the julia command from your terminal emulator. For MacOS, you'll need to add Julia to your path. See this StackOverflow post if you're having issues.

Add this package by adding the github url. First, enter package mode ]. Then, type add https://github.com/graphprotocol/AllocationOpt.jl. You'll also want to add the Comonicon package.

pkg> add https://github.com/graphprotocol/AllocationOpt.jl/
pkg> add Comonicon

Public API

We expose the following functions in our public API.

AllocationOpt.network_state — Method

function network_state(id, whitelist, blacklist, pinnedlist, frozenlist, indexer_service_network_url)

Arguments

id::AbstractString: The id of the indexer to optimise.

whitelist::Vector{AbstractString}: Subgraph deployment IPFS hashes included in this list will be considered for, but not guaranteed allocation.

blacklist::Vector{AbstractString}: Subgraph deployment IPFS hashes included in this list will not be considered, and will be suggested to close if there's an existing allocation.

pinnedlist::Vector{AbstractString}: Subgraph deployment IPFS hashes included in this list will be guaranteed allocation. Currently unsupported.

frozenlist::Vector{AbstractString}: Subgraph deployment IPFS hashes included in this list will not be considered during optimisation. Any allocations you have on these subgraphs deployments will remain.

indexer_service_network_url::AbstractString: The URL that exposes the indexer service's network endpoint. Must begin with http. Example: http://localhost:7600/network.

AllocationOpt.optimize_indexer — Method

function optimize_indexer(indexer, repo, minimum_allocation_amount, maximum_new_allocations)

Arguments

indexer::Indexer: The indexer being optimised.

repo::Repository: Contains the current network state.

minimum_allocation_amount::Real: The minimum amount of GRT that you are willing to allocate to a subgraph.

maximum_new_allocations::Integer: The maximum number of new allocations you would like the optimizer to open.

```

AllocationOpt.push_allocations! — Method

function push_allocations!(indexer_id, management_server_url, proposed_allocations, whitelist, blacklist, pinnedlist, frozenlist)

Arguments

indexer_id::AbstractString: Indexer id

management_server_url::T: Indexer management server url, in format similar to http://localhost:18000

proposed_allocations::Dict{T,<:Real}: The set of allocation to open returned by optimize_indexer

whitelist::AbstractVector{T}: Unused. Here for completeness.

blacklist::AbstractVector{T}: Unused. Here for completeness.

pinnedlist::AbstractVector{T}: Unused. Here for completeness.

frozenlist::AbstractVector{T}: Make sure to not close these allocations as they are frozen.

AllocationOpt.read_filterlists — Method

function read_filterlists(filepath)

Arguments

filepath::AbstractString: A path to the CSV file that contains whitelist, blacklist, pinnedlist, frozenlist as columns.